+++ kitty/options/types.py
@@ -53,6 +53,7 @@ option_names = (
     'allow_cloning',
     'allow_hyperlinks',
     'allow_remote_control',
+    'auto_reload_config',
     'background',
     'background_blur',
     'background_image',
@@ -527,6 +528,7 @@ class Options:
     allow_cloning: choices_for_allow_cloning = 'ask'
     allow_hyperlinks: int = 1
     allow_remote_control: choices_for_allow_remote_control = 'no'
+    auto_reload_config: float = 0.1
     background: Color = Color(0, 0, 0)
     background_blur: int = 0
     background_image: tuple[str, ...] = ()